Diego Mateos: fixes (no peta al salir, puntero lista msg, quats...) + updates usan double + log jsons

+++ Proyecto3/src/GameLogic/AutoRotationComponent.cpp
@@ -4,7 +4,7 @@ void AutoRotationComponent::setUp()
 {
 }
 
-void AutoRotationComponent::update(GameObject * ent, float time)
+void AutoRotationComponent::update(GameObject * ent, double time)
 {
 	ent->setOrientation(RenderSystemInterface::getSingleton()->getRotationFrom_To(ent->getPosition(), dest));
 }
