+++ Proyecto3/src/GameLogic/GameObject.cpp
@@ -10,7 +10,7 @@ GameObject::GameObject(nap_json const & cfg, std::list<Component*> comps)
 
 GameObject::~GameObject() {
 	for (auto comp : components_) {
-		if (comp != nullptr)delete comp;
+		if (comp != nullptr) delete comp;
 	}
 
 	components_.clear();
@@ -24,13 +24,10 @@ void GameObject::setUp() {
 	if (cfg_.find("pos") != cfg_.end()) setPosition(nap_vector3(cfg_["pos"]));
 	if (cfg_.find("ori") != cfg_.end()) setOrientation(nap_quat(cfg_["ori"]));
 
-
-
 	//init components
 	for (auto comp : components_) if (comp != nullptr) comp->setUp();
 
 	//send user ptr to physics component (done by component factory)
-
 	//getComponent("Phys")->receive(&Msg_PX_userPtr(getTransPtr()));
 }
 
@@ -52,11 +49,11 @@ bool GameObject::handleEvents(const SDL_Event evt) {
 	return handled;
 }
 
-void GameObject::update(float time) {
+void GameObject::update(double time) {
 	for (auto comp : components_)
 		if (comp != nullptr && comp->isActive())comp->update(this, time);
 }
-void GameObject::late_update(float time) {
+void GameObject::late_update(double time) {
 	for (auto comp : components_)
 		if (comp != nullptr && comp->isActive())comp->late_update(this, time);
 }
